Added util and tests. Renamed handler to controller. Extracted limit to config. Miscelanious small code changes.

+++ model/tournament.go
@@ -15,7 +15,7 @@ type TournamentData struct {
 }
 
 type Tournament struct {
-	ID   int    `json:"_id"`
+	ID   int64  `json:"_id"`
 	Name string `json:"name"`
 	Year string `json:"year"`
 }
